LPS-98560 Create new factory to handle toolbar display context, depending on the type. Move the existing toolbar display context to basic type
